i got bug in when one time signin after , react-google-login login onsuccess method auto execute when page refresh

+++ src/component/0Auth/OAuthLogin.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import {GoogleLogin} from 'react-google-login'
 import { CLIENT_ID } from '../../../env'
 import styled, { css } from "styled-components";
@@ -7,6 +7,7 @@ import './Styles.css'
 import { useDispatch } from 'react-redux';
 import { deleteProfileAction, setProfileAction } from '../../StatesManagement/reducers/UserSigninReducerslice';
 import { useNavigate } from 'react-router-dom';
+import { toast } from "react-toastify";
 // import dotenv from 'dotenv'
 // dotenv.config() 
 
@@ -14,18 +15,43 @@ import { useNavigate } from 'react-router-dom';
 const OAuthLogin = () => {
   const dispatch =  useDispatch()
   const navigate = useNavigate()
+  const [alreadySignIn,setAlreadySignIn] = useState(false)
 
     const onSuccess =(res)=>{
-      if(res && localStorage.getItem("active_user" === false)){
+      console.log("succes",res)
+      if(res && localStorage.getItem("active_user") == "false" && alreadySignIn){
         localStorage.setItem("active_user",true)
         dispatch(setProfileAction(res))
-        console.log("/todo")
+        toast.success('Success Full', {
+          position: "top-right",
+          autoClose: 5000,
+          hideProgressBar: false,
+          closeOnClick: true,
+          pauseOnHover: true,
+          draggable: true,
+          progress: undefined,
+          theme: "light",
+          });
+        
+        navigate("/todo")
       }
+      setAlreadySignIn(false)
     }
 
     const onFailure =(res)=>{
       dispatch(deleteProfileAction(res))
       localStorage.setItem("active_user",false)
+      toast.error('Error', {
+        position: "top-right",
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        theme: "light",
+        });
+        navigate("/login")
     }
